+++ src/utils/logIn&SignUp/setUserData.ts
@@ -1,0 +1,10 @@
+import { setDoc } from "firebase/firestore";
+import { db } from "../../Firebase";
+import { doc } from "firebase/firestore";
+import { assetData } from "./useAuthorization";
+
+const setUserData = async (userUid: string) => {
+  await setDoc(doc(db, "user", userUid), assetData);
+};
+
+export default setUserData;
